% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitStMoMo.R
\name{applyIdentifiabilyConstraints}
\alias{applyIdentifiabilyConstraints}
\title{Apply the identfiability constraints to a model}
\usage{
applyIdentifiabilyConstraints(object, fittedCoef, oxt, wxt, ages, years)
}
\arguments{
\item{object}{an object of class \code{"StMoMo"} defining the stochastic
mortality model.}

\item{fittedCoef}{a list with current estimas of ax, bx, kt, b0x, gc}

\item{oxt}{matrix/vector or scalar of known offset used in fitting
the model.}

\item{wxt}{matrix of 0-1 weights to be used in the fitting process.}

\item{ages}{ages in the fitting data.}

\item{years}{years in the fitting data.}
}
\value{
a list of parameter estimates with th identifiability constraints applied
}
\description{
Applies the identifiability constraints and checks that they preserve the link
}
\keyword{internal}
