% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalUtils.R
\name{arima.string}
\alias{arima.string}
\title{Copy of unexported function arima.string from forecast}
\usage{
arima.string(object, padding = FALSE)
}
\description{
Copy of unexported function arima.string from forecast
}
\keyword{internal}

