% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapfitStMoMo.R
\name{binomRes2q}
\alias{binomRes2q}
\title{Map Binomial deviance residuals into deaths}
\usage{
binomRes2q(dhat, e, res)
}
\arguments{
\item{dhat}{fitted number of deaths (scalar)}

\item{e}{number of exposures (scalar)}

\item{res}{deviance residual (scalar)}
}
\value{
matching observed deaths
}
\description{
This funciont uses the procedure described in 
Debon et al. (2010, section 3)
}
\references{
Debon, A., Martinez-Ruiz, F., & Montes, F. (2010). A geostatistical 
approach for dynamic life tables: The effect of mortality on remaining 
lifetime and annuities. Insurance: Mathematics and Economics, 
47(3), 327-336.
}
\keyword{internal}

