% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportUtils.R
\name{central2initial}
\alias{central2initial}
\title{Transform StMoMoData from central to initial exposures}
\usage{
central2initial(data)
}
\arguments{
\item{data}{StMoMoData object of type "central" created with function 
\code{\link{StMoMoData}}.}
}
\value{
A StMoMoData object of type "initial".
}
\description{
Transform StMoMoData from central to initial exposures. Initial exposures
are computed by adding one half of the deaths to the central exposures.
}
\seealso{
\code{\link{initial2central}}
}
