% Generated by roxygen2 (4.0.2): do not edit by hand
\name{computeDevianceBinomial}
\alias{computeDevianceBinomial}
\title{Binomial deviance}
\usage{
computeDevianceBinomial(obs, fit, exposure, weight)
}
\arguments{
\item{obs}{observed rates}

\item{fit}{fitted rates}

\item{exposure}{observed exposure}

\item{weight}{weigths given to each observation#'}
}
\description{
Binomial deviance
}
\keyword{internal}

