% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalUtils.R
\name{computeDevianceGaussian}
\alias{computeDevianceGaussian}
\title{Gaussian deviance}
\usage{
computeDevianceGaussian(obs, fit, weight)
}
\arguments{
\item{obs}{observed log rates}

\item{fit}{fitted log rates}

\item{weight}{weights given to each observation}

\item{sigma2}{estimated vairance}
}
\description{
Gaussian deviance
}
\keyword{internal}
