% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalUtils.R
\name{computeLogLikGaussian}
\alias{computeLogLikGaussian}
\title{Compute Gaussian loglikelihod}
\usage{
computeLogLikGaussian(obs, fit, weight)
}
\arguments{
\item{obs}{observed log rates}

\item{fit}{fitted log rates}

\item{weight}{weights given to each observation}

\item{sigma2}{estimated vairance}
}
\description{
Compute Gaussian loglikelihod
}
\keyword{internal}
