% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/internalUtils.R
\name{computeLogLikPoisson}
\alias{computeLogLikPoisson}
\title{Compute Poisson loglikelihod}
\usage{
computeLogLikPoisson(obs, fit, weight)
}
\arguments{
\item{obs}{observed number of deaths}

\item{fit}{fitted numberd of deaths}

\item{weight}{weigths given to each observation#'}
}
\description{
Compute Poisson loglikelihod
}
\keyword{internal}

