% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvStMoMo.R
\name{cvStMoMo}
\alias{cvStMoMo}
\title{Cross validation for Stochastic Mortality Model}
\usage{
cvStMoMo(object, h = NULL, data = NULL, Dxt = NULL, Ext = NULL,
  ages.train = NULL, years.train = NULL, ages = NULL, years = NULL,
  returnY = FALSE, type = c("logrates", "rates"), verbose = TRUE)
}
\arguments{
\item{object}{an object of class \code{"StMoMo"} defining the 
stochastic mortality model.}

\item{h}{number of years for forecasting horizon.}

\item{data}{an optional object of type StMoMoData containing 
information on deaths and exposures to be used for training the model. 
This is typically created with function \code{\link{StMoMoData}}. 
If this is not provided then the training data is taken from 
arguments, \code{Dxt}, \code{Ext}, \code{ages}, \code{years}.}

\item{Dxt}{optional matrix of deaths data.}

\item{Ext}{optional matrix of observed exposures of the same 
dimension of \code{Dxt}.}

\item{ages.train}{optional vector of ages to include in the 
training. Must be a subset of \code{ages}.}

\item{years.train}{optional vector of years to include in the 
training. Must be a subset of \code{years}.}

\item{ages}{optional vector of ages corresponding to rows of 
\code{Dxt} and \code{Ext}.}

\item{years}{optional vector of years corresponding to rows of 
\code{Dxt} and \code{Ext}.}

\item{returnY}{a logical value. If \code{TRUE}, \code{cvStMoMo}
returns the fitted mortality rates for the cross validation
folds.}

\item{type}{the type of the predicted values that the cross 
validation is performed with respect to. The alternatives are 
\code{"rates"} and \code{"logrates"}. If \code{"rates"}, \code{cvStMoMo}
returns the mean squared error using the deviation between
realised and predicted mortality rates. If \code{"logrates"}, 
\code{cvStMoMo} returns the mean squared error using the 
deviation between realised and predicted log mortality rates.}

\item{verbose}{a logical value. If \code{TRUE} progress 
indicators are printed as the model is fitted. Set 
\code{verbose = FALSE} to silent the fitting and avoid 
progress messages.}
}
\value{
A list with class \code{"cvStMoMo"} with components:
  
  \item{model}{ the object of class \code{"StMoMo"} defining 
  the fitted stochastic mortality model.}
  
  \item{data}{ StMoMoData object provided for training the model.}
  
  \item{Dxt}{ matrix of deaths used in the training.}
  
  \item{Ext}{ matrix of exposures used in the training.}
  
  \item{rates}{ matrix of mortality rates (\eqn{D_{x,t}/E_{x,t}}) 
  used in the training.} 
   
  \item{logrates}{ matrix of log mortality rates (\eqn{log(D_{x,t}/E_{x,t})}) 
  used in the training.} 
  
  \item{cv.rates}{ if \code{returnY=TRUE}, matrix of predicted 
  mortality rates.} 
  
  \item{cv.mse}{ if \code{type="rates"}, mean squared error using 
  predicted mortality rates and realised mortality rates. If 
  \code{type="logrates"}, mean squared error using predicted log 
  mortality rates and realised log mortality rates.}
}
\description{
Perform cross validation by period for a Stochastic Mortality Model.
}
\examples{

# Lee-Carter model only for one year forecasting horizon based on log-rates
LCcv <- cvStMoMo(lc(), h = 1, data = EWMaleData, ages.train = 55:89, type = "logrates")

# APC model using arguments Dxt, Ext, ages, years to pass fitting data, based on rates
APCcv <- cvStMoMo(apc(), h = 10, Dxt = EWMaleData$Dxt, Ext = EWMaleData$Ext, 
ages = EWMaleData$ages, years = EWMaleData$years, ages.train = 55:89, type = "rates")
            
}
