% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractStMoMo.R
\name{extractStMoMo}
\alias{extractStMoMo}
\title{Extract a Stochastic Mortality Model fit from group regularised fitting}
\usage{
extractStMoMo(object, k, simplify = TRUE)
}
\arguments{
\item{object}{an object of class \code{"grpfitStMoMo"} with the regularisation
path of the fitted Stochastic Mortality Model}

\item{k}{integer index indicating which position (lambda) in the regularisation 
path is being extracted}

\item{simplify}{optional logical value indicating if the model should be simplified
to only include model terms which are different from 0.}
}
\value{
An list of class \code{"fitStMoMo"} with a fitted Stochastic Mortality Model.
See \code{\link{fit.StMoMo}} for the details of the elements in the list.
}
\description{
Extract a Stochastic Mortality Model fit from group regularised fitting.
This is useful to get the fit for a single lambda in the regularisation path
of a group regularised fit obtained with function \code{\link{grpfit}}. 
The output is a \code{"fitStMoMo"} object which can then be used with other 
StMoMo functions such as \code{\link{plot.fitStMoMo}}, 
\code{\link{forecast.fitStMoMo}} and \code{\link{simulate.fitStMoMo}}
}
\examples{
APCgrpfit <- grpfit(apc(link = "log-Gaussian"), data = EWMaleData, 
                    ages.fit = 20:89, years.fit = 1965:2010) 
APCgrp <- extractStMoMo(APCgrpfit, 15)                     
plot(APCgrp, parametricbx = FALSE)
plot(forecast(APCgrp), parametricbx = FALSE)

}
