% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrwd.R
\name{forecast}
\alias{forecast}
\title{Generic for forecasting a Stochastic Mortality Model or Random-Walk with drift model.}
\usage{
forecast(object, ...)
}
\arguments{
\item{object}{an object used to select a method. Typically of class 
\code{StMoMo} or an extension of this class.}

\item{...}{arguments to be passed to or from other methods.}
}
\description{
Generic for forecasting a Stochastic Mortality Model or Random-Walk with drift model.
}
\keyword{internal}
