% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genFunctions.R
\name{genAbove}
\alias{genAbove}
\title{Generate above indicator functions}
\usage{
genAbove(strikes)
}
\arguments{
\item{strikes}{vector of strikes.}
}
\value{
A list of above indicator functions for the given vector of strikes.
}
\description{
Generates a list of above indicator functions given a vector of strikes.
This is useful for constructing a large abstract Stochastic
Mortality Model to be fitted using the group lasso.
}
