% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genFunctions.R
\name{genBelow}
\alias{genBelow}
\title{Generate below indicator functions}
\usage{
genBelow(strikes)
}
\arguments{
\item{strikes}{vector of strikes.}
}
\value{
A list of below indicator functions for the given vector of strikes.
}
\description{
Generates a list of above indicator functions given a vector of strikes.
This is useful for constructing a large abstract Stochastic
Mortality Model to be fitted using the group lasso.
}
