% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapfitStMoMo.R
\name{genBinomialResBootSamples}
\alias{genBinomialResBootSamples}
\title{Generate Binomial residual bootstrap sample}
\usage{
genBinomialResBootSamples(devRes, dhat, E, nBoot)
}
\arguments{
\item{devRes}{Matrix of deviance residuals}

\item{dhat}{Matrix of fitted deaths}

\item{E}{Matrix of observed exposures}

\item{nBoot}{number of bootstrap samples to produce.}
}
\value{
a list of length \code{nBoot} of matrices with matching sampled 
deaths
}
\description{
Generate Binomial bootstrap samples using the procedure
described in Debon et al. (2010, section 3)
}
\references{
Debon, A., Martinez-Ruiz, F., & Montes, F. (2010). A geostatistical 
approach for dynamic life tables: The effect of mortality on remaining 
lifetime and annuities. Insurance: Mathematics and Economics, 47(3), 
327-336.
}
\keyword{internal}
