% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genFunctions.R
\name{genCallSq}
\alias{genCallSq}
\title{Generate call squared functions}
\usage{
genCallSq(strikes)
}
\arguments{
\item{strikes}{vector of strikes.}
}
\value{
A list of call squared functions for the given vector of strikes.
}
\description{
Generates a list of call squared functions given a vector of strikes.
For each strike \eqn{k}, the functions are of the form
\deqn{f(x) = \max(x-k,0)^2} 
This is useful for constructing a large abstract Stochastic
Mortality Model to be fitted using the group lasso.
}
