% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapfitStMoMo.R
\name{genPoissonResBootSamples}
\alias{genPoissonResBootSamples}
\title{Generate poisson residual bootstrap sample}
\usage{
genPoissonResBootSamples(devRes, dhat, nBoot)
}
\arguments{
\item{devRes}{Matrix of deviance residuals}

\item{dhat}{Marix of fitted deaths}

\item{nBoot}{number of bootstrap samples to produce.}
}
\value{
a list of length \code{nBoot} of marices with matching sampled 
deaths
}
\description{
Generate poisson bootstrap samples using the procedure
described in Renshaw and Haberman (2008)
}
\references{
Renshaw, A. E., & Haberman, S. (2008). On simulation-based approaches to 
risk measurement in mortality with specific reference to Poisson Lee-Carter 
modelling. Insurance: Mathematics and Economics, 42(2), 797-816.
}
\keyword{internal}

