% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapfitStMoMo.R
\name{genPoissonSemiparametricBootSamples}
\alias{genPoissonSemiparametricBootSamples}
\title{Generate Poisson semiparametric bootstrap samples}
\usage{
genPoissonSemiparametricBootSamples(D, nBoot)
}
\arguments{
\item{D}{matrix of deaths}

\item{nBoot}{number of bootstrap samples to produce.}
}
\value{
a list of length \code{nBoot} of matrices with sampled deaths
}
\description{
Generate Poisson semiparametric bootstrap samples using the procedure
described in Brouhns et al (2005).
}
\references{
Brouhns, N., Denuit M., & Van Keilegom, I. (2005). Bootstrapping the 
Poisson log-bilinear model for mortality forecasting. 
Scandinavian Actuarial Journal, 2005(3), 212-224.
}
\keyword{internal}
