% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genFunctions.R
\name{genPoly}
\alias{genPoly}
\title{Generate polynomial functions}
\usage{
genPoly(exp)
}
\arguments{
\item{exp}{vector of exponents.}
}
\value{
A list of polynomial functions for the given vector of exponents.
}
\description{
Generates a list of polynomial functions given a vector of exponents.
This is useful for constructing a large abstract Stochastic
Mortality Model to be fitted using the group lasso.
}
