% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genFunctions.R
\name{genPutSq}
\alias{genPutSq}
\title{Generate put squared functions}
\usage{
genPutSq(strikes)
}
\arguments{
\item{strikes}{vector of strikes.}
}
\value{
A list of put squared functions for the given vector of strikes.
}
\description{
Generates a list of put squared functions given a vector of strikes.
For each strike \eqn{k}, the functions are of the form
\deqn{f(x) = \max(k-x,0)^2} 
This is useful for constructing a large abstract Stochastic
Mortality Model to be fitted using the group lasso.
}
