% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glRemoveIntercept.R
\name{glRemoveIntercept}
\alias{glRemoveIntercept}
\title{Remove Intercept from Coefficients}
\usage{
glRemoveIntercept(object, tempfittedCoef, intercept, ages)
}
\arguments{
\item{object}{an object of class \code{"StMoMo"} defining the stochastic
mortality model, with only parametric age-period and age-cohort terms.}

\item{intercept}{estimated intercept from \code{grpreg}.}

\item{ages}{vector of ages corresponding to rows of \code{Dxt} and \code{Ext}.}

\item{tempfitttedCoef}{estimated coefficients from \code{grpreg} using
\code{extractCoefficientsFromGnm}.}
}
\description{
Remove Intercept from Coefficients
}
\keyword{internal}
