% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exportUtils.R
\name{initial2central}
\alias{initial2central}
\title{Transform StMoMoData from initial to central exposures}
\usage{
initial2central(data)
}
\arguments{
\item{data}{StMoMoData object of type "initial" created with function
\code{\link{StMoMoData}}.}
}
\value{
A StMoMoData object of type "central".
}
\description{
Transform StMoMoData from initial to central exposures. Central exposures
are computed by substracting one half of the deaths from the initial exposures.
}
\seealso{
\code{\link{central2initial}}
}

