% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/internalUtils.R
\name{logit}
\alias{logit}
\title{Logit function
\code{logit} computes the logit function}
\usage{
logit(x)
}
\description{
Logit function
\code{logit} computes the logit function
}
\keyword{internal}

