% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logLikfitStMoMo.R
\name{logLik.fitStMoMo}
\alias{logLik.fitStMoMo}
\title{Log-Likelihood of a fitStMoMo object}
\usage{
\method{logLik}{fitStMoMo}(object, ...)
}
\arguments{
\item{object}{an object of class \code{fitStMoMo} representing a
Stochastic Mortality Model fitted to some data.}

\item{...}{other arguments.}
}
\value{
The log-likelihood of the fitted model.
}
\description{
Returns the log-likelihood of a fitted Stochastic Mortality Model.
}

