% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLikfitGL.R
\name{logLik.grpfitStMoMo}
\alias{logLik.grpfitStMoMo}
\title{Log-Likelihood of a fitGL object}
\usage{
\method{logLik}{grpfitStMoMo}(object, ...)
}
\arguments{
\item{object}{an object of class \code{fitGL} representing a 
Stochastic Mortality Model fitted to some data using the group lasso function
\code{\link{glStMoMo}}.}

\item{...}{other arguments.}
}
\value{
The log-likelihood of the fitted model.
}
\description{
Returns the log-likelihood of a  Stochastic Mortality Model fitted with 
group regularised penalty.
}
