% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnm_Mult.R
\name{Mult}
\alias{Mult}
\title{Hack gnm::Mult weird dependency
The way this function is called in StMoMo functions makes it necesary 
to have the gnm dependecy. A workaround is redefine it in the package and 
export it from here, and import the gnm package.}
\source{
\code{\link[gnm]{Mult}} or from here: 
\url{https://github.com/hturner/gnm/blob/master/R/Mult.R}
}
\usage{
Mult(..., inst = NULL)
}
\arguments{
\item{...}{a comma-separated list of two or more symbolic
    expressions representing the constituent multipliers in the
    interaction.}

\item{inst}{a positive integer specifying the instance number of the
    term.}
}
\description{
Hack gnm::Mult weird dependency
The way this function is called in StMoMo functions makes it necesary 
to have the gnm dependecy. A workaround is redefine it in the package and 
export it from here, and import the gnm package.
}
\keyword{internal}
