% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardModels.R
\name{plat}
\alias{plat}
\title{Create a Plat type mortality model}
\usage{
plat(link = c("log", "logit", "log-Gaussian"), simplified = FALSE)
}
\arguments{
\item{link}{defines the link function and random component associated with 
the mortality model. \code{"log"} would assume that deaths follow a 
Poisson distribution and use a log link.  \code{"logit"} would assume 
that deaths follow a Binomial distribution and a logit link. 
\code{"log-Gaussian"} would assume that log death rates follow a 
normal (Gaussian) distribution and an identity link.}

\item{simplified}{logical value indicating whether the \eqn{\kappa_t^{(3)}}
should be included in the model. Default is \code{TRUE} so this term is included. 
If \code{TRUE} this term is excluded and we obtaind the so-called simplified Plat model
which is useful when only older age are of interst.}
}
\value{
An object of class \code{"StMoMo"}.
}
\description{
Utility function to initialise a \code{StMoMo} object representing the 
mortality model proposed by Plat (2009).
}
\details{
The created model is either a logit-Binomial, a log-Poisson or log-Gaussian
version of the Plat model which has predictor structure 
\deqn{\eta_{xt} = \alpha_x+ \kappa_t^{(1)} + (\bar{x}-x)\kappa_t^{(2)} + (\bar{x}-x)^+\kappa_t^{(3)} + \gamma_{t-x}}
where \eqn{\bar{x}} is the average age in the data and 
\eqn{(\bar{x}-x)^+=\max(\bar{x}-x, 0)}. 

Identifiability of the model is accomplished by applying parameters 
constraints \eqn{\sum_t\kappa__t^{(1)} = \sum_t\kappa__t^{(2)} = \sum_t\kappa__t^{(2)} = 0} 
and \deqn{\sum_c\gamma_c = 0, \sum_c c\gamma_c = 0, 
\sum_c c^2\gamma_c = 0} which ensure that the cohort effect fluctuates 
around zero and has no linear or quadratic trend
}
\examples{

PLAT <- plat()
wxt <- genWeightMat(ages = 0:89, years = EWMaleData$years, clip = 3)
PLATfit <- fit(PLAT, data = EWMaleData, ages.fit = 0:89, wxt = wxt)
plot(PLATfit, parametricbx = FALSE)

sPLAT <- plat(simplified = TRUE)
wxt <- genWeightMat(ages = 55:89, years = EWMaleData$years, clip = 3)
sPLATfit <- fit(sPLAT, data = EWMaleData, ages.fit = 55:89, wxt = wxt)
plot(sPLATfit, parametricbx = FALSE)


}
\references{
Plat, R. (2009). On Stochastic Mortality Modelling. 
Insurance: Mathematics and Economics, 43(3), 393-404.
}
\seealso{
\code{\link{StMoMo}}
}
