% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glcvStMoMo.R
\name{plot.cv.grpStMoMo}
\alias{plot.cv.grpStMoMo}
\title{Plot the cross-validation curve produced by cv.grpreg}
\usage{
\method{plot}{cv.grpStMoMo}(x, sign.lambda = 1, ...)
}
\arguments{
\item{x}{an object of class \code{"cv.grpStMoMo"} with the cross-validation results
of the group regularised fitting  of a stochastic mortality model.}

\item{sign.lambda}{Either plot against \code{log(lambda)} (default) or its
negative if \code{sign.lambda=-1}.}

\item{...}{additional arguments to control graphical appearance.
See \code{\link[graphics]{plot}}.}
}
\description{
Plots the cross-validation currve, upper and lower stamdard errpr curves,
as a function of the \code{lambda} values used.
}
\seealso{
\code{cv.grpStMoMo}.
}
