% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-grpfitStMoMo.R
\name{plot.grpfitStMoMo}
\alias{plot.grpfitStMoMo}
\title{Plot the regularisation path of a Stochasic Mortality Model}
\usage{
\method{plot}{grpfitStMoMo}(x, label = FALSE, ...)
}
\arguments{
\item{x}{an object of class \code{"grpfitStMoMo"}.}

\item{label}{If TRUE, annotates the plot with text labels in the right margin 
describing which age/period or cohort terms the lines belong o}

\item{...}{additional arguments to control graphical appearance.
See \code{\link[graphics]{plot}}.}
}
\description{
Plots the regularisation path of a Stochasic Mortality Model 
fitted with function \code{grpfit}.
}
