% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotbootStMoMo.R
\name{plotParameterFan}
\alias{plotParameterFan}
\title{Plot fanchart of the parameters}
\usage{
plotParameterFan(x, y, yBoot, main, xlab, ylab, probs, fan.col, n.fan, ...)
}
\arguments{
\item{x}{value of x}

\item{y}{central value of y}

\item{yBoot}{matrix with the simulated values of y}

\item{main}{title of the plot}

\item{xlab}{label of x axis}

\item{ylab}{label of y axis}

\item{probs}{probabilities related to percentiles to plot in the fan chart}

\item{fan.col}{palette of colours used in the fan chart}

\item{n.fan}{the number of colour to use in the fan}
}
\description{
Plot fanchart of the parameters
}
\details{
In order for the plotting to look appropiately the intervals of 
the data with non missing values need to be found. Otherwise fanplot::fan 
doesn't work.
}
\keyword{internal}
