% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapfitStMoMo.R
\name{poissonRes2death}
\alias{poissonRes2death}
\title{Map poisson deviance residuals into deaths}
\usage{
poissonRes2death(dhat, res)
}
\arguments{
\item{dhat}{fitted number of deaths (scalar)}

\item{res}{deviance residual (scalar)}
}
\value{
matching observed deaths
}
\description{
This function uses the procedure described in Renshaw and Haberman (2008)
}
\references{
Renshaw, A. E., & Haberman, S. (2008). On simulation-based approaches to 
risk measurement in mortality with specific reference to Poisson Lee-Carter 
modelling. Insurance: Mathematics and Economics, 42(2), 797-816.
}
\keyword{internal}
