% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalUtils.R
\name{predictLink}
\alias{predictLink}
\title{Compute the link for a given mortality models}
\usage{
predictLink(ax, bx, kt, b0x, gc, oxt, ages, years)
}
\description{
Compute the link for a given mortality models
}
\keyword{internal}

