% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapfitStMoMo.R
\name{print.bootStMoMo}
\alias{print.bootStMoMo}
\title{Print an object of class \code{"bootStMoMo"}}
\usage{
\method{print}{bootStMoMo}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"bootStMoMo"}.}

\item{...}{arguments to be passed to or from other methods.}
}
\description{
\code{print} method for class \code{"bootStMoMo"}.
}

