% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitStMoMo.R
\name{print.fitStMoMo}
\alias{print.fitStMoMo}
\title{Print an object of class \code{"fitStMoMo"}}
\usage{
\method{print}{fitStMoMo}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"fitStMoMo"}.}

\item{...}{arguments to be passed to or from other methods.}
}
\description{
\code{print} method for class \code{"fitStMoMo"}.
}

