% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/forecastfitStMoMo.R
\name{print.forStMoMo}
\alias{print.forStMoMo}
\title{Print an object of class \code{"forStMoMo"}}
\usage{
\method{print}{forStMoMo}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"forStMoMo"}.}

\item{...}{arguments to be passed to or from other methods.}
}
\description{
\code{print} method for class \code{"forStMoMo"}.
}

