% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulatefitStMoMo.R
\name{print.simStMoMo}
\alias{print.simStMoMo}
\title{Print an object of class \code{"simStMoMo"}}
\usage{
\method{print}{simStMoMo}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"simStMoMo"}.}

\item{...}{arguments to be passed to or from other methods.}
}
\description{
\code{print} method for class \code{"simStMoMo"}.
}

