% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/StMoMo.R
\name{print.StMoMo}
\alias{print.StMoMo}
\title{Print an object of class \code{"StMoMo"}}
\usage{
\method{print}{StMoMo}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"StMoMo"}.}

\item{...}{arguments to be passed to or from other methods.}
}
\description{
\code{print} method for class \code{"StMoMo"}.
}
\details{
\code{print.StMoMo} prints a description of the Stochastic Mortality
Model
}

