% Generated by roxygen2 (4.0.2): do not edit by hand
\name{residuals.fitStMoMo}
\alias{residuals.fitStMoMo}
\title{Extract deviance residuals of a Stochastic
Mortality Model}
\usage{
\method{residuals}{fitStMoMo}(object, scale = TRUE, ...)
}
\arguments{
\item{object}{an object of class \code{"fitStMoMo"} with the fitted
parameters of a stochastic mortality model.}

\item{scale}{logical indicating scaling or not the residual by dividing the deviance by the
overdispersion of the model.  Default is \code{TRUE}.}

\item{...}{other arguments.}
}
\value{
An object of class \code{"resStMoMo"} with the residuals. This object
has components:
  \item{residuals}{ a matrix with the residuals.}
  \item{ages}{ ages corresponding to the rows in \code{residuals}.}
  \item{years}{ years corresponding to the columns in \code{residuals}.}
}
\description{
Compute deviance residuals of a fitted Stochastic Mortality Model.
}
\examples{
CBDfit <- fit(cbd(), Dxt = EWMaleData$Dxt, Ext = EWMaleData$Ext,
              ages = EWMaleData$ages, years = EWMaleData$years,
              ages.fit = 55:89)
CBDres <- residuals(CBDfit)
plot(CBDres)
}

