% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/residualsfitStMoMo.R
\name{scatterplotAPC}
\alias{scatterplotAPC}
\title{Do a scatter plot of a matrix according to age-period-cohorts}
\usage{
scatterplotAPC(mat, ages, years, plotAge = TRUE, plotYear = TRUE,
  plotCohort = TRUE, zeroLine = TRUE, ...)
}
\arguments{
\item{mat}{matrix with the data to plot.}

\item{ages}{ages corresponding to the rows in \code{mat}.}

\item{years}{years corresponding to the columns in \code{mat}.}

\item{plotAge}{logical value indicating if the age scatter plot should be
produced.}

\item{plotYear}{logical value indicating if the calendar year scatter plot
should be produced.}

\item{plotCohort}{logical value indicating if the cohort scatter plot
should be produced.}

\item{zeroLine}{logical valae indicating if a horizontal line at zero
should be plotted.}

\item{...}{other arguments to pass to the plot function.}
}
\description{
Do a scatter plot of a matrix according to age-period-cohorts
}
\keyword{internal}

