% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iarima.R
\name{simulate.iarima}
\alias{simulate.iarima}
\title{Simulate independent arima series}
\usage{
\method{simulate}{iarima}(object, nsim = 10, seed = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{"iarima"}.}

\item{nsim}{number of periods for the simulated series.}

\item{seed}{either \code{NULL} or an integer that will be used in a 
call to \code{\link{set.seed}} before simulating the time series. 
The default, \code{NULL} will not change the random generator state.}

\item{...}{other arguments.}
}
\description{
Returns one simulated path of the group of independent 
arima series in \code{object}.
}
