% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mrwd.R
\name{simulate.mrwd}
\alias{simulate.mrwd}
\title{Simulate a Multivariate Random Walk with Drift}
\usage{
\method{simulate}{mrwd}(object, nsim = 10, seed = NULL, ...)
}
\arguments{
\item{object}{An object of class "mrwd".}

\item{nsim}{number of periods for the simulated series.}

\item{seed}{either \code{NULL} or an integer that will be used in a call to
\code{\link{set.seed}} before simulating the time series.
The default, \code{NULL} will not change the random generator state.}

\item{...}{other arguments.}
}
\description{
Returns a  simulated multivariate time series based on the Multivariate
Random Walk with Drift model in \code{object}.
}

