% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalUtils.R
\name{computeDevianceBinomial}
\alias{computeDevianceBinomial}
\title{Binomial deviance}
\usage{
computeDevianceBinomial(obs, fit, exposure, weight)
}
\arguments{
\item{obs}{observed rates}

\item{fit}{fitted rates}

\item{exposure}{observed exposure}

\item{weight}{weights given to each observation#'}
}
\description{
Binomial deviance
}
\keyword{internal}
