% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalUtils.R
\name{computeDeviancePoisson}
\alias{computeDeviancePoisson}
\title{Compute Poisson deviance}
\usage{
computeDeviancePoisson(obs, fit, weight)
}
\arguments{
\item{obs}{observed number of deaths}

\item{fit}{fitted numberd of deaths}

\item{weight}{weigths given to each observation#'}
}
\description{
Compute Poisson deviance
}
\keyword{internal}
