% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internalUtils.R
\name{computeLogLikBinomial}
\alias{computeLogLikBinomial}
\title{Compute Binomial loglikelihod}
\usage{
computeLogLikBinomial(obs, fit, exposure, weight)
}
\arguments{
\item{obs}{observed rates}

\item{fit}{fitted rates}

\item{exposure}{observed exposure}

\item{weight}{weigths given to each observation#'}
}
\description{
Compute Binomial loglikelihod
}
\keyword{internal}

