% adfagaGenerated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{EWMaleData}
\alias{EWMaleData}
\title{England and Wales male mortality data}
\format{A list  with the following components:
\describe{
  \item{Dxt}{ matrix of deathd data.}
  \item{Ext}{ matrix of exposured data (mid year population estimates).}
  \item{ages}{ vector of ages.}
  \item{years}{ vector of years.}  
}}
\source{
Human Mortality Database \url{http://www.mortality.org/}.
}
\usage{
EWMaleData
}
\description{
Age-specific deaths and exposures for England and Wales from
the Human Mortality Database.
}
\details{
\code{EWMaleData} contains deaths and exposures for England and 
Wales males for the period 1961-2011 and for ages 0-100.
Data taken from the Human Mortality Database on 5 November 2014.
}
\references{
Human Mortality Database (2014). University of California,
Berkeley (USA), and Max Planck Institute for Demographic Research (Germany).
Available at \url{www.mortality.org}.
}
\keyword{datasets}

