% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fitStMoMo.R
\name{fit}
\alias{fit}
\title{Generic for fitting a Stochastic Mortality Model}
\usage{
fit(object, ...)
}
\arguments{
\item{object}{an object used to select a method. Typically of class \code{StMoMo} or
an extension of this class.}

\item{...}{arguments to be passed to or from other methods.}
}
\description{
\code{fit} is a generic function for fitting Stochastic Mortality Models.
The function invokes particular methods which depend on the class of the first argument.
}
\details{
\code{fit} is a generic function which means that new fitting strategies can be added
for particular stochastic mortality models. See for instance \code{\link{fit.StMoMo}}.
}

