% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genFunctions.R
\name{genCall}
\alias{genCall}
\title{Generate call functions}
\usage{
genCall(strikes)
}
\arguments{
\item{strikes}{vector of strikes.}
}
\value{
A list of call functions for the given vector of strikes.
}
\description{
Generates a list of call functions given a vector of strikes.
This is useful for constructing a large abstract Stochastic
Mortality Model to be fitted using the group lasso.
}
