% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glStMoMo.R
\name{grpfit}
\alias{grpfit}
\title{Fit a Stochastic Mortality Model with Group Regularised Penalties}
\usage{
grpfit(object, lambda = NULL, nlambda = 50, data = NULL,
  Dxt = NULL, Ext = NULL, ages = NULL, years = NULL,
  ages.fit = NULL, years.fit = NULL, wxt = NULL, index = NULL,
  verbose = TRUE)
}
\arguments{
\item{object}{an object of class \code{"StMoMo"} defining the stochastic
mortality model, with only parametric age-period and age-cohort terms.}

\item{lambda}{a user supplied sequence of lambda values. If left unspecified,
the function will compute a grid of lambda values of length \code{nlambda} 
that range uniformly on the log scale.}

\item{nlambda}{number of lambda values. Default is 50.}

\item{data}{an optional object of type StMoMoData containing information on
deaths and exposures to be used for fitting the model. This is typically created 
with  function \code{\link{StMoMoData}}. If this is not provided then the fitting 
data is taken from arguments, \code{Dxt}, \code{Ext}, \code{ages}, \code{years}.}

\item{Dxt}{optional matrix of deaths data.}

\item{Ext}{optional matrix of observed exposures of the same dimension of 
\code{Dxt}.}

\item{ages}{optional vector of ages corresponding to rows of \code{Dxt} and 
\code{Ext}.}

\item{years}{optional vector of years corresponding to rows of \code{Dxt} and 
\code{Ext}.}

\item{ages.fit}{optional vector of ages to include in the fit. Must be a 
subset of \code{ages}.}

\item{years.fit}{optional vector of years to include in the fit. Must be a 
subset of \code{years}.}

\item{wxt}{optional matrix of 0-1 weights to be used in the fitting process. 
This can be used, for instance, to zero weight some cohorts in the data.
See \code{\link{genWeightMat}} which is a helper function for defining 
weighting matrices.}

\item{index}{optional vector describing the grouping of the coefficients. If
there are coefficients to be included in the model without being penalised, assign
them to group 0 (or "0"). If this is not provided, the grouping is automatically
determined and all groups are penalised.}

\item{verbose}{a logical value. If \code{TRUE} progress indicators are 
printed as the model is fitted. Set \code{verbose = FALSE} to silent the 
fitting and avoid progress messages.}
}
\value{
A list with class \code{"grpfitStMoMo"} with components:
  
  \item{model}{ the object of class \code{"StMoMo"} defining the fitted 
  stochastic mortality model.}
  
  \item{beta}{ list of fitted coefficients for each value of lambda. Each
  entry contains \code{ax}, \code{kt}, and \code{gc}}. 
  
  \item{ax}{ vector with the fitted values of the static age function 
  \eqn{\alpha_x}. If the model does not have a static age function or 
  failed to fit this is set to \code{NULL}.}
  
  \item{kt}{ matrix with the values of the fitted period indexes 
  \eqn{\kappa_t^{(i)}, i=1, ..., N}. \code{kt[i, ]} contains the estimated 
  values of the \eqn{i}-th period index. If the model does not have any 
  age-period terms (i.e. \eqn{N=0}) or failed to fit this is set to 
  \code{NULL}.}
    
  \item{gc}{ vector with the fitted cohort index \eqn{\gamma_{c}}.
  If the model does not have a cohort effect or failed to fit this is set
  to \code{NULL}.}
  
  \item{data}{ StMoMoData object provided for fitting the model.}
  
  \item{Dxt}{ matrix of deaths used in the fitting.}
  
  \item{Ext}{ matrix of exposures used in the fitting.}
  
  \item{wxt}{ matrix of 0-1 weights used in the fitting.}
  
  \item{ages}{ vector of ages used in the fitting.}
  
  \item{years}{ vector of years used in the fitting.}
  
  \item{cohorts}{ vector of cohorts used in the fitting.}
  
  \item{loglik}{ vector of log-likelihoods of fitted model.}
  
  \item{lambda}{ vector of lambda values used in the fitting.}
}
\description{
Fit a Stochastic Mortality Model with Group Regularised Penalties
to a given data set. The fitting is done using package \code{grpreg}.
}
\details{
Fitting is done using function \code{\link[grpreg]{grpreg}} within package 
\code{grpreg}. This is achieved by minimising the residual sum of squares subject 
to the minimax concave penalty (MCP). This function is only supported for 
log-Gaussian models and is not yet supported for logit-Binomial or log-Poisson 
models. Ages and years in the data should be of type numeric. Data points with 
zero exposure are assigned a zero weight and are ignored in the fitting process. 
Similarly, \code{NA} are assigned a zero weight and ignored in the fitting process.
}
\examples{
APCgrpfit <- grpfit(apc(link = "log-Gaussian"), data = EWMaleData, 
                    ages.fit = 20:89, years.fit = 1965:2010) 
plot(extractStMoMo(APCgrpfit, 5), parametricbx = FALSE, nCol = 3)
plot(extractStMoMo(APCgrpfit, 15), parametricbx = FALSE, nCol = 3)
plot(extractStMoMo(APCgrpfit, 20), parametricbx = FALSE, nCol = 3)                                                                                                

}
