% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalUtils.R
\name{invlogit}
\alias{invlogit}
\title{Inverse Logit function
\code{invlogit} computes the inverse logit function}
\usage{
invlogit(x)
}
\description{
Inverse Logit function
\code{invlogit} computes the inverse logit function
}
\keyword{internal}
