% Generated by roxygen2 (4.0.2): do not edit by hand
\name{invlogit}
\alias{invlogit}
\title{Invers Logit function
\code{invlogit} computes the inverse logit function}
\usage{
invlogit(x)
}
\description{
Invers Logit function
\code{invlogit} computes the inverse logit function
}
\keyword{internal}

