% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glcvStMoMo.R
\name{plot.cv.grpStMoMo}
\alias{plot.cv.grpStMoMo}
\title{Plot the cross-validation curve produced by cv.grpreg}
\usage{
\method{plot}{cv.grpStMoMo}(x, sign.lambda = 1, ...)
}
\arguments{
\item{x}{an object of class \code{"cv.grpStMoMo"} with the cross-validation results
of the group regularised fitting  of a stochastic mortality model.}

\item{sign.lambda}{Either plot against \code{log(lambda)} (default) or its
negative if \code{sign.lambda=-1}.}

\item{...}{additional arguments to control graphical appearance.
See \code{\link[graphics]{plot}}.}
}
\description{
Plots the cross-validation currve, upper and lower stamdard errpr curves,
as a function of the \code{lambda} values used.
}
\examples{


APCcvgrp <- cv.grpStMoMo(apc(link = "log-Gaussian"), h = 1, data = EWMaleData, 
                         ages.train = 55:89, nlambda = 25)

plot(APCcvgrp)


#Long computing times
\dontrun{
#Create big model
strikes <- seq(25,85,5)
bModel <- StMoMo(link = "log-Gaussian", staticAgeFun = TRUE, 
                 periodAgeFun = c("1", genPoly(2:10), genCall(strikes),
                                                   genPut(strikes)),
                 cohortAgeFun = "1")
                 
#Cross-validation for 20 year ahead forecast                  
EWcv.20 <- cv.grpStMoMo(bModel, h = 20, data = EWMaleData, 
                        ages.train = 20:89)
                        
plot(EWcv.20)
}                       


}
\seealso{
\code{cv.grpStMoMo}.
}
